+++ huggle/wikiuser.hpp
@@ -135,7 +135,7 @@ namespace Huggle
             //! may be wrong, don't relly on it
             bool IsReported;
             //! Number of contributions, if it's not known, it contains negative value
-            int EditCount;
+            long EditCount;
             //! This is a mediawiki string containing a time when user was registered
             QString RegistrationDate;
             //! Groups that this user is in, by default it's empty
